Deplace LoadDetailXmlSerializer in iidm-xml-converter
